examples(bank-manager): replace strawman demo with two realistic-baseline behaviors
Removes the old three-beat demo whose baseline agent had no authorization
logic at all, and replaces it with two behaviors that each start from a
baseline a reviewer would sign off on.

- Behavior 1 (sensitivity-tier authorization): deterministic decision,
  property-based Rego rule vs a competent-but-domain-scoped Python gate.
- Behavior 2 (coercion via unverified authority): non-deterministic
  judgement, calibrated classifier annotator vs a control-aware prompt
  plus keyword tripwire.

agent.py is renamed to bank_agent_common.py and stripped to shared
plumbing only (LLM construction, MCP server startup, text extraction);
it no longer declares a system prompt or an ASSERT callable.

Also fixes a pre-existing pytest collection failure in the example's
tests/ package via tests/conftest.py.
