+++ examples/bank_manager_agent_control/.env.example
@@ -19,7 +19,7 @@ AZURE_API_KEY=<paste-your-azure-openai-key>
 AZURE_API_BASE=https://<your-azure-openai-resource>.openai.azure.com/
 AZURE_API_VERSION=2024-12-01-preview
 # Deployment name on your resource. MUST match the azure/<name> model strings
-# in the eval_realistic_*.yaml configs (currently azure/gpt-5.4-mini). If your
+# in the eval_tier_authorization*.yaml and eval_coercion_*.yaml configs. If your
 # deployment is named differently, change BOTH here and in the YAMLs so the
 # agent-under-test and the eval/judge use the same unfiltered deployment.
 AGENT_MODEL=gpt-5.4-mini
