+++ talks/README.md
@@ -7,7 +7,8 @@ viewer. No browser, build step, or sibling folders required.
 |---|---|---|
 | **AI Engineer World's Fair 2026** — *evaluate and optimize your AI agents* | [`aiewf-18min/aiewf-2026-deck.pdf`](aiewf-18min/aiewf-2026-deck.pdf) | 18 min · 7 slides |
 
-The AIEWF deck walks the same three beats as the
-[`bank_manager_agent_control`](../examples/bank_manager_agent_control/README.md) example:
-baseline → defensive prompting → principled control plane, then the ASSERT + ACS
-announcement.
+The AIEWF deck walks a three-beat arc — baseline → defensive prompting → principled
+control plane — then the ASSERT + ACS announcement. The
+[`bank_manager_agent_control`](../examples/bank_manager_agent_control/README.md) example
+now runs that same arc as **two** behaviors, each against a realistic baseline rather than
+an unguarded agent; see the example README for the current arms and results.
